+++ src/engine/tm/entities/AnimatedPlayer.java
@@ -53,10 +53,10 @@ public class AnimatedPlayer extends AnimatedModel implements IPlayer {
 			super.decreasePosition(dx, upwardsSpeed, dz);
 		}
 
-		ITerrain currentTerrain = scene.getCurrentTerrain(super.getPosition().x, super.getPosition().z);
+		ITerrain currentTerrain = scene.getCurrentTerrain((int) super.getPosition().x, (int) super.getPosition().z);
 		float terrainHeight = TERRAIN_HEIGHT;
 		if (currentTerrain != null) {
-			terrainHeight = currentTerrain.getHeightOfTerrain(super.getPosition().x, super.getPosition().z);
+			terrainHeight = currentTerrain.getHeightOfTerrain((int) super.getPosition().x, (int) super.getPosition().z);
 		}
 
 		if (super.getPosition().y < terrainHeight) {
