+++ src/engine/tm/entities/Camera.java
@@ -119,9 +119,9 @@ public class Camera implements ICamera {
 		position.y = player.getPosition().y + verticalDistance + offsetY;
 		position.z = player.getPosition().z - offsetZ;
 		
-		ITerrain terrain = scene.getCurrentTerrain(position.x, position.z);
-		if (terrain instanceof ITerrain && terrain.getHeightOfTerrain(position.x, position.z) > position.y) {
-			position.y = terrain.getHeightOfTerrain(position.x, position.z) + 5;
+		ITerrain terrain = scene.getCurrentTerrain((int) position.x, (int) position.z);
+		if (terrain instanceof ITerrain && terrain.getHeightOfTerrain((int) position.x, (int) position.z) > position.y) {
+			position.y = terrain.getHeightOfTerrain((int) position.x, (int) position.z) + 5;
 		}
 	}
